+++ internal/app/tokenised-infrastructure-rewarder/types/response_types.go
@@ -51,7 +51,7 @@ type FarmHashRateElement struct {
 	Time             string  `json:"time"`
 	UnixTime         int64   `json:"unixTime"`
 	AverageWorkers   float64 `json:"averageWorkers"`
-	HashrateAccepted int64   `json:"hashrateAccepted"`
+	HashrateAccepted float64 `json:"hashrateAccepted"`
 	HashrateStale    int64   `json:"hashrateStale"`
 	HashrateRejected int64   `json:"hashrateRejected"`
 }
